+++ go-client-sdk/sdk.go
@@ -85,8 +85,8 @@ func WithClient(client HTTPClient) SDKOption {
 func New(opts ...SDKOption) *SDK {
 	sdk := &SDK{
 		_language:   "go",
-		_sdkVersion: "0.5.0",
-		_genVersion: "2.29.0",
+		_sdkVersion: "0.6.0",
+		_genVersion: "2.30.0",
 	}
 	for _, opt := range opts {
 		opt(sdk)