+++ python-client-sdk/src/sdk/contextid.py
@@ -28,11 +28,11 @@ class ContextID:
         base_url = self._server_url
         
         url = utils.generate_url(operations.AddRuleRequest, base_url, '/{contextID}/rules/add', request)
-        
         headers = {}
         req_content_type, data, form = utils.serialize_request_body(request, "add_rule_input", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
+        headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
         
@@ -54,11 +54,11 @@ class ContextID:
         base_url = self._server_url
         
         url = utils.generate_url(operations.EvaluateRulesRequest, base_url, '/{contextID}/evaluate', request)
-        
         headers = {}
         req_content_type, data, form = utils.serialize_request_body(request, "evaluate_rules_input", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
+        headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
         
@@ -80,11 +80,12 @@ class ContextID:
         base_url = self._server_url
         
         url = utils.generate_url(operations.GetAllRulesRequest, base_url, '/{contextID}/rules', request)
-        
+        headers = {}
+        headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
         
-        http_res = client.request('GET', url)
+        http_res = client.request('GET', url, headers=headers)
         content_type = http_res.headers.get('Content-Type')
 
         res = operations.GetAllRulesResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
@@ -102,11 +103,12 @@ class ContextID:
         base_url = self._server_url
         
         url = utils.generate_url(operations.RemoveRuleRequest, base_url, '/{contextID}/rules/{id}', request)
-        
+        headers = {}
+        headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)
         
-        http_res = client.request('DELETE', url)
+        http_res = client.request('DELETE', url, headers=headers)
         content_type = http_res.headers.get('Content-Type')
 
         res = operations.RemoveRuleResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
@@ -120,11 +122,11 @@ class ContextID:
         base_url = self._server_url
         
         url = utils.generate_url(operations.ReplaceRuleRequest, base_url, '/{contextID}/rules/{id}', request)
-        
         headers = {}
         req_content_type, data, form = utils.serialize_request_body(request, "replace_rule_input", 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
+        headers['user-agent'] = f'speakeasy-sdk/{self._language} {self._sdk_version} {self._gen_version}'
         
         client = utils.configure_security_client(self._client, security)